getrecords when new record is added
